move ServiceUtilities to topl-service-base module
