+++ topl-service-base/src/main/java/io/matthewnelson/topl_service_base/ServiceUtilities.kt
@@ -64,7 +64,7 @@
 *     modified version of TorOnionProxyLibrary-Android, and you must remove this
 *     exception when you distribute your modified version.
 * */
-package io.matthewnelson.topl_service.util
+package io.matthewnelson.topl_service_base
 
 import android.annotation.SuppressLint
 import java.text.NumberFormat
@@ -84,6 +84,7 @@ object ServiceUtilities {
 
     @SuppressLint("ConstantLocale")
     private val numberFormat = NumberFormat.getInstance(Locale.getDefault())
+
     /**
      * Obtained from: https://gitweb.torproject.org/tor-android-service.git/tree/service/
      *                src/main/java/org/torproject/android/service/TorEventHandler.java
@@ -94,11 +95,11 @@ object ServiceUtilities {
         return if (value < 1e6) {
             numberFormat.format(
                 Math.round((((value * 10 / 1024).toInt()) / 10).toFloat())
-            ) + "kbps"
+            ) + "KBps"
         } else {
             numberFormat.format(
                 Math.round((((value * 100 / 1024 / 1024).toInt()) / 100).toFloat())
-            ) + "mbps"
+            ) + "MBps"
         }
     }
 }
